attendance-system/backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
httpx>=0.27.0
python-dotenv>=1.0.0